+++ packages/ripple-tide-grant/src/index.vue
@@ -46,11 +46,6 @@ export default { name: 'TideGrantPage' }
 
 <script setup lang="ts">
 import type { TideGrantPage } from './types'
-import TideGrantHeader from './components/tide-grant-header.vue'
-import TideGrantOverview from './components/tide-grant-overview.vue'
-import TideGrantTimeline from './components/tide-grant-timeline.vue'
-import TideGrantGuidelines from './components/tide-grant-guidelines.vue'
-import TideGrantDocuments from './components/tide-grant-documents.vue'
 
 interface Props {
   site: any