+++ packages/ripple-tide-publication/src/components/tide-publication-header.vue
@@ -3,7 +3,7 @@ export default { name: 'TidePublicationHeader' }
 </script>
 
 <template>
-  <RplHeroHeader
+  <RplHeaderHeroHeader
     :corner-top="true"
     :corner-bottom="true"
     :behind-nav="true"
@@ -11,12 +11,11 @@ export default { name: 'TidePublicationHeader' }
     :title="header.title"
   >
     <p class="rpl-type-p-large">{{ header.summary }}</p>
-  </RplHeroHeader>
+  </RplHeaderHeroHeader>
 </template>
 
 <script setup lang="ts">
 import type { TidePublicationHeader } from '../types'
-import { RplHeroHeader } from '@dpc-sdp/ripple-ui-core'
 
 defineProps<{
   header: TidePublicationHeader