+++ packages/ripple-tide-publication/src/types.ts
@@ -14,11 +14,30 @@ export type TidePublicationChapter = {
   link: any
 }
 
+export interface apiNode {
+  title: string
+  url: string
+  id: string
+  children: apiNode[] | undefined
+}
+
+export interface indexNode {
+  text: string
+  url: string
+  id: string
+  items: indexNode[] | undefined
+  active: boolean | undefined
+}
+
 export interface TidePublicationPage extends TidePageBase {
   /**
    * @description Page title
    */
   title: string
+  url: string
+  background: string
+  showInPageNav: boolean
+  inPageNavHeadingLevel: string
   /**
    * @description RplHeader
    */
@@ -38,5 +57,9 @@ export interface TidePublicationPage extends TidePageBase {
   /**
    * @description Page action documents
    */
-  documents: any
+  // documents: any
+
+  publication: Array<any>
+  // children: Array<any>
+  breadcrumbs: Array<any>
 }